Struct parry2d::bounding_volume::AABB

pub struct AABB {
    pub mins: Point<Real>,
    pub maxs: Point<Real>,
}

An Axis Aligned Bounding Box.

Fields
mins: Point<Real>maxs: Point<Real>
Implementations

impl AABB

pub fn new(mins: Point<Real>, maxs: Point<Real>) -> AABB

Creates a new AABB.

Arguments:

mins - position of the point with the smallest coordinates.

maxs - position of the point with the highest coordinates. Each component of mins must be smaller than the related components of maxs.

pub fn new_invalid() -> Self

Creates an invalid AABB with mins components set to Real::max_values and maxscomponents set to -Real::max_values.

This is often used as the initial values of some AABB merging algorithms.

pub fn from_half_extents(
center: Point<Real>,
half_extents: Vector<Real>
) -> Self

Creates a new AABB from its center and its half-extents.

pub fn from_points<'a, I>(pts: I) -> Self where
I: IntoIterator<Item = &'a Point<Real>>,

Creates a new AABB from a set of points.

pub fn center(&self) -> Point<Real>

The center of this AABB.

pub fn half_extents(&self) -> Vector<Real>

The half extents of this AABB.

pub fn extents(&self) -> Vector<Real>

The extents of this AABB.

pub fn take_point(&mut self, pt: Point<Real>)

Enlarges this AABB so it also contains the point pt.

pub fn transform_by(&self, m: &Isometry<Real>) -> Self

Computes the AABB bounding self transformed by m.

pub fn bounding_sphere(&self) -> BoundingSphere

The smallest bounding sphere containing this AABB.

pub fn contains_local_point(&self, point: &Point<Real>) -> bool

pub fn vertices(&self) -> [Point<Real>; 4]

Computes the vertices of this AABB.

pub fn split_at_center(&self) -> [AABB; 4]

Splits this AABB at its center, into four parts (as in a quad-tree).

impl AABB

pub fn clip_segment(
&self,
pa: &Point<Real>,
pb: &Point<Real>
) -> Option<Segment>

Computes the intersection of a segment with this AABB.

Returns None if there is no intersection.

pub fn clip_line_parameters(
&self,
orig: &Point<Real>,
dir: &Vector<Real>
) -> Option<(Real, Real)>

Computes the parameters of the two intersection points between a line and this AABB.

The parameters are such that the point are given by orig + dir * parameter. Returns None if there is no intersection.

pub fn clip_line(
&self,
orig: &Point<Real>,
dir: &Vector<Real>
) -> Option<Segment>

Computes the intersection segment between a line and this AABB.

Returns None if there is no intersection.

pub fn clip_ray_parameters(&self, ray: &Ray) -> Option<(Real, Real)>

Computes the parameters of the two intersection points between a ray and this AABB.

The parameters are such that the point are given by ray.orig + ray.dir * parameter. Returns None if there is no intersection.

pub fn clip_ray(&self, ray: &Ray) -> Option<Segment>

Computes the intersection segment between a ray and this AABB.

Returns None if there is no intersection.

impl AABB

pub fn clip_polygon(&self, points: &mut Vec<Point<Real>>)

Computes the intersections between this AABB and the given polygon.

The results is written into points directly. The input points are assumed to form a convex polygon where all points lie on the same plane. In order to avoid internal allocations, uses self.clip_polygon_with_workspace instead.

pub fn clip_polygon_with_workspace(
&self,
points: &mut Vec<Point<Real>>,
workspace: &mut Vec<Point<Real>>
)

Computes the intersections between this AABB and the given polygon.

The results is written into points directly. The input points are assumed to form a convex polygon where all points lie on the same plane.