Module parry2d::math

Compilation flags dependent aliases for mathematical types.

Constants

DEFAULT_EPSILON
The default tolerance used for geometric operations.

DIM
The dimension of the space.

SIMD_LAST_INDEX
SIMD_WIDTH - 1

SIMD_WIDTH
The number of lanes of a SIMD number.

Type Definitions

AngVector
The angular vector type.

AngularInertia
The angular inertia of a rigid body.

CrossMatrix
A matrix that represent the cross product with a given vector.

Dim
The dimension of the ambient space.

Isometry
The transformation matrix type.

Matrix
The matrix type.

Orientation
The orientation type.

Point
The point type.

PrincipalAngularInertia
The principal angular inertia of a rigid body.

Real
The scalar type used throughout this crate.

Rotation
The rotation matrix type.

SdpMatrix
A 2D symmetric-definite-positive matrix.

SimdBool
A SIMD bool with SIMD_WIDTH lanes.

SimdReal
A SIMD float with SIMD_WIDTH lanes.

SpacialVector
A vector with a dimension equal to the maximum number of degrees of freedom of a rigid body.

Translation
The translation type.

UnitVector
The unit vector type.

Vector
The vector type.