Struct parry2d::partitioning::QBVH

pub struct QBVH<T> { /* fields omitted */ }

A quaternary bounding-volume-hierarchy.

This is a bounding-volume-hierarchy where each node has either four children or none.

Implementations

impl<T: IndexedData> QBVH<T>

pub fn new() -> Self

Initialize an empty quad-tree.

pub fn root_aabb(&self) -> &AABB

The AABB of the root of this tree.

pub fn leaf_data(&mut self, node_id: NodeIndex) -> Option<T>

Returns the data associated to a given leaf.

Returns None if the provided node ID does not identify a leaf.

pub fn clear_and_rebuild(
&mut self,
data_gen: impl QBVHDataGenerator<T>,
dilation_factor: Real
)

Clears this quad-tree and rebuilds it from a new set of data and AABBs.

pub fn pre_update(&mut self, data: T)

Marks a piece of data as dirty so it can be updated during the next call to self.update.

pub fn update<F>(&mut self, aabb_builder: F, dilation_factor: Real) where
F: Fn(&T) -> AABB,

Update all the nodes that have been marked as dirty by self.pre_update.

pub fn intersect_aabb(&self, aabb: &AABB, out: &mut Vec<T>)

Retrieve all the data of the nodes with AABBs intersecting the given AABB:

pub fn traverse_depth_first(&self, visitor: &mut impl SimdVisitor<T, SimdAABB>)

Performs a depth-first traversal on the BVH.

pub fn traverse_depth_first_with_stack(
&self,
visitor: &mut impl SimdVisitor<T, SimdAABB>,
stack: &mut Vec<u32>
)

Performs a depth-first traversal on the BVH.

pub fn traverse_best_first<BFS>(
&self,
visitor: &mut BFS
) -> Option<(NodeIndex, BFS::Result)> where
BFS: SimdBestFirstVisitor<T, SimdAABB>,
BFS::Result: Clone,

Performs a best-first-search on the BVH.

Returns the content of the leaf with the smallest associated cost, and a result of user-defined type.