Trait parry2d::partitioning::QBVHDataGenerator

pub trait QBVHDataGenerator<T> {
    fn size_hint(&self) -> usize;

    fn for_each(&mut self, f: impl FnMut(T, AABB));
}

Trait used for generating the content of the leaves of the QBVH acceleration structure.

Required methods

fn size_hint(&self) -> usize

Gives an idea of the number of elements this generator contains.

This is primarily used for pre-allocating some arrays for better performances.

fn for_each(&mut self, f: impl FnMut(T, AABB))

Iterate through all the elements of this generator.

Loading content...