parry2d::partitioning
Trait QbvhDataGenerator

pub trait QbvhDataGenerator<LeafData> {
    // Required methods
    fn size_hint(&self) -> usize;
    fn for_each(&mut self, f: impl FnMut(LeafData, Aabb));
}

Trait used for generating the content of the leaves of the Qbvh acceleration structure.

Required Methods

fn size_hint(&self) -> usize

Gives an idea of the number of elements this generator contains.

This is primarily used for pre-allocating some arrays for better performances.

fn for_each(&mut self, f: impl FnMut(LeafData, Aabb))

Iterate through all the elements of this generator.

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.