Module parry2d::query::contact

Implementation details of the contact and contacts functions.

Structs

Contact
Geometric description of a contact.

Functions

contact
Computes one pair of contact points point between two shapes.

contact_ball_ball
Contact between balls.

contact_ball_convex_polyhedron
Contact between a ball and a convex polyhedron.

contact_composite_shape_shape
Best contact between a composite shape (Mesh, Compound) and any other shape.

contact_convex_polyhedron_ball
Contact between a convex polyhedron and a ball.

contact_cuboid_cuboid
Contact between two cuboids.

contact_halfspace_support_map
Contact between a halfspace and a support-mapped shape (Cuboid, ConvexHull, etc.)

contact_shape_composite_shape
Best contact between a shape and a composite (Mesh, Compound) shape.

contact_support_map_halfspace
Contact between a support-mapped shape (Cuboid, ConvexHull, etc.) and a halfspace.

contact_support_map_support_map
Contact between support-mapped shapes (Cuboid, ConvexHull, etc.)

contact_support_map_support_map_with_params
Contact between support-mapped shapes (Cuboid, ConvexHull, etc.)