parry2d::query::details
Function contact_manifold_convex_ball_shapes

pub fn contact_manifold_convex_ball_shapes<ManifoldData, ContactData>(
    pos12: &Isometry<Real>,
    shape1: &dyn Shape,
    shape2: &dyn Shape,
    prediction: Real,
    manifold: &mut ContactManifold<ManifoldData, ContactData>,
)
where
    ContactData: Default + Copy,

Computes the contact manifold between a convex shape and a ball, both represented as a Shape trait-object.