Module parry2d::query::sat

Application of the Separating-Axis-Theorem (SAT).

Functions

cuboid_cuboid_find_local_separating_normal_oneway

Finds the best separating normal between two cuboids.

cuboid_support_map_find_local_separating_normal_oneway

Finds the best separating normal between a cuboid and a convex shape implementing the SupportMap trait.

point_cuboid_find_local_separating_normal_oneway

Computes the separation between a point and a cuboid, along the given direction normal1.

segment_cuboid_find_local_separating_normal_oneway

Finds the best separating normal between a cuboid and a segment.

support_map_support_map_compute_separation

Computes the separation along the given direction, between two convex shapes implementing the SupportMap trait.

triangle_cuboid_find_local_separating_normal_oneway

Finds the best separating normal between a triangle and a cuboid.

triangle_support_map_find_local_separating_normal_oneway

Finds the best separating normal between a triangle and a convex shape implementing the SupportMap trait.