Trait parry2d::query::PersistentQueryDispatcher

pub trait PersistentQueryDispatcher<ManifoldData = (), ContactData = ()>: QueryDispatcher {
    fn contact_manifolds(
        &self, 
        pos12: &Isometry<Real>, 
        g1: &dyn Shape, 
        g2: &dyn Shape, 
        prediction: Real, 
        manifolds: &mut Vec<ContactManifold<ManifoldData, ContactData>>, 
        workspace: &mut Option<ContactManifoldsWorkspace>
    ) -> Result<(), Unsupported>;

    fn contact_manifold_convex_convex(
        &self, 
        pos12: &Isometry<Real>, 
        g1: &dyn Shape, 
        g2: &dyn Shape, 
        prediction: Real, 
        manifold: &mut ContactManifold<ManifoldData, ContactData>
    ) -> Result<(), Unsupported>;
}

A query dispatcher for queries relying on spatial coherence, including contact-manifold computation.

Required methods

fn contact_manifolds(
&self,
pos12: &Isometry<Real>,
g1: &dyn Shape,
g2: &dyn Shape,
prediction: Real,
manifolds: &mut Vec<ContactManifold<ManifoldData, ContactData>>,
workspace: &mut Option<ContactManifoldsWorkspace>
) -> Result<(), Unsupported>

Compute all the contacts between two shapes.

The output is written into manifolds and context. Both can persist between multiple calls to contacts by re-using the result of the previous call to contacts. This persistence can significantly improve collision detection performances by allowing the underlying algorithms to exploit spatial and temporal coherence.

fn contact_manifold_convex_convex(
&self,
pos12: &Isometry<Real>,
g1: &dyn Shape,
g2: &dyn Shape,
prediction: Real,
manifold: &mut ContactManifold<ManifoldData, ContactData>
) -> Result<(), Unsupported>

Computes the contact-manifold between two convex shapes.

Loading content...