parry2d::query
Trait WorkspaceData

pub trait WorkspaceData: DowncastSync {
    // Required methods
    fn as_typed_workspace_data(&self) -> TypedWorkspaceData<'_>;
    fn clone_dyn(&self) -> Box<dyn WorkspaceData>;
}

Data from a ContactManifoldsWorkspace.

Required Methods

fn as_typed_workspace_data(&self) -> TypedWorkspaceData<'_>

Gets the underlying workspace as an enum.

fn clone_dyn(&self) -> Box<dyn WorkspaceData>

Clones self.

Implementations

impl dyn WorkspaceData

pub fn is<__T: WorkspaceData>(&self) -> bool

Returns true if the trait object wraps an object of type __T.

pub fn downcast<__T: WorkspaceData>( self: Box<Self>, ) -> Result<Box<__T>, Box<Self>>

Returns a boxed object from a boxed trait object if the underlying object is of type __T. Returns the original boxed trait if it isn’t.

pub fn downcast_rc<__T: WorkspaceData>( self: Rc<Self>, ) -> Result<Rc<__T>, Rc<Self>>

Returns an Rc-ed object from an Rc-ed trait object if the underlying object is of type __T. Returns the original Rc-ed trait if it isn’t.

pub fn downcast_ref<__T: WorkspaceData>(&self) -> Option<&__T>

Returns a reference to the object within the trait object if it is of type __T, or None if it isn’t.

pub fn downcast_mut<__T: WorkspaceData>(&mut self) -> Option<&mut __T>

Returns a mutable reference to the object within the trait object if it is of type __T, or None if it isn’t.

pub fn downcast_arc<__T: WorkspaceData + Any + Send + Sync>( self: Arc<Self>, ) -> Result<Arc<__T>, Arc<Self>>

Returns an Arc-ed object from an Arc-ed trait object if the underlying object is of type __T. Returns the original Arc-ed trait if it isn’t.