parry2d::query
Module visitors

Visitors for performing geometric queries exploiting spatial partitioning data structures.

Structs

AabbSetsInterferencesCollector

Spatial partitioning data structure visitor collecting interferences with a given bounding volume.

BoundingVolumeIntersectionsSimultaneousVisitor

Spatial partitioning data structure visitor collecting interferences with a given bounding volume.

BoundingVolumeIntersectionsVisitor

Spatial partitioning data structure visitor collecting interferences with a given bounding volume.

CompositeClosestPointVisitor

Best-first traversal visitor for computing the point closest to a composite shape.

CompositePointContainmentTest

Visitor for checking if a composite shape contains a specific point.

PointIntersectionsVisitor

Spatial partitioning structure visitor collecting nodes that may contain a given point.

RayIntersectionsVisitor

Bounding Volume Tree visitor collecting intersections with a given ray.