Module parry2d::shape

Shapes supported by parry.

Structs

Ball
A Ball shape.

Capsule
A capsule shape defined as a round segment.

Compound
A compound shape with an aabb bounding volume.

ConvexPolygon
A 2D convex polygon.

Cuboid
Shape of a box.

HalfSpace
A half-space delimited by an infinite plane.

HeightField
A 2D heightfield.

PolygonalFeature
A polygonal feature representing the local polygonal approximation of a vertex, or face, of a convex shape.

Polyline
A polyline.

RoundShape
A shape with rounded borders.

Segment
A segment shape.

SharedShape
The shape of a collider.

TriMesh
A triangle mesh.

Triangle
A triangle shape.

Enums

FeatureId
An identifier of a feature of a convex polyhedron.

SegmentPointLocation
Logical description of the location of a point on a triangle.

ShapeType
Enum representing the type of a shape.

TrianglePointLocation
Description of the location of a point on a triangle.

TypedShape
Enum representing the shape with its actual type

Traits

PolygonalFeatureMap
Trait implemented by convex shapes with features with polyhedral approximations.

Shape
Trait implemented by shapes usable by Rapier.

SimdCompositeShape
Trait implemented by shapes composed of multiple simpler shapes.

SupportMap
Traits of convex shapes representable by a support mapping function.

TypedSimdCompositeShape

Type Definitions

RoundConvexPolygon
A convex polygon dilated by a sphere (so it has round corners).

RoundCuboid
A cuboid dilated by a sphere (so it has round corners).

RoundTriangle
A triangle dilated by a sphere (so it has round corners).