Struct parry2d::shape::ConvexPolygon

pub struct ConvexPolygon { /* fields omitted */ }

A 2D convex polygon.

Implementations

impl ConvexPolygon

pub fn aabb(&self, pos: &Isometry<Real>) -> AABB

Computes the world-space AABB of this convex polygon, transformed by pos.

pub fn local_aabb(&self) -> AABB

Computes the local-space AABB of this convex polygon.

impl ConvexPolygon

pub fn bounding_sphere(&self, pos: &Isometry<Real>) -> BoundingSphere

Computes the world-space bounding sphere of this convex polygon, transformed by pos.

pub fn local_bounding_sphere(&self) -> BoundingSphere

Computes the local-space bounding sphere of this convex polygon.

impl ConvexPolygon

pub fn from_convex_hull(points: &[Point<Real>]) -> Option<Self>

Creates a new 2D convex polygon from an arbitrary set of points.

This explicitly computes the convex hull of the given set of points. Use Returns None if the convex hull computation failed.

pub fn from_convex_polyline(points: Vec<Point<Real>>) -> Option<Self>

Creates a new 2D convex polygon from a set of points assumed to describe a counter-clockwise convex polyline.

Convexity of the input polyline is not checked. Returns None if all points form an almost flat line.

pub fn points(&self) -> &[Point<Real>]

The vertices of this convex polygon.

pub fn normals(&self) -> &[Unit<Vector<Real>>]

The normals of the edges of this convex polygon.

pub fn support_feature_id_toward(
&self,
local_dir: &Unit<Vector<Real>>
) -> FeatureId

Get the ID of the feature with a normal that maximizes the dot product with local_dir.

pub fn feature_normal(&self, feature: FeatureId) -> Option<Unit<Vector<Real>>>

The normal of the given feature.