parry2d::shape
Struct TriMesh

#[repr(C)]
pub struct TriMesh { /* private fields */ }

A triangle mesh.

Implementations

impl TriMesh

pub fn bounding_sphere(&self, pos: &Isometry<f32>) -> BoundingSphere

Computes the world-space bounding sphere of this triangle mesh, transformed by pos.

pub fn local_bounding_sphere(&self) -> BoundingSphere

Computes the local-space bounding sphere of this triangle mesh.

impl TriMesh

pub fn new(vertices: Vec<Point<f32>>, indices: Vec<[u32; 3]>) -> Self

Creates a new triangle mesh from a vertex buffer and an index buffer.

pub fn with_flags( vertices: Vec<Point<f32>>, indices: Vec<[u32; 3]>, flags: TriMeshFlags, ) -> Self

Creates a new triangle mesh from a vertex buffer and an index buffer, and flags controlling optional properties.

pub fn set_flags(&mut self, flags: TriMeshFlags) -> Result<(), TopologyError>

Sets the flags of this triangle mesh, controlling its optional associated data.

pub fn transform_vertices(&mut self, transform: &Isometry<f32>)

Transforms in-place the vertices of this triangle mesh.

pub fn scaled(self, scale: &Vector<f32>) -> Self

Returns a scaled version of this triangle mesh.

pub fn append(&mut self, rhs: &TriMesh)

Appends a second triangle mesh to this triangle mesh.

pub fn from_polygon(vertices: Vec<Point<f32>>) -> Option<Self>

Create a TriMesh from a set of points assumed to describe a counter-clockwise non-convex polygon.

This operation may fail if the input polygon is invalid, e.g. it is non-simple or has zero surface area.

pub fn flat_indices(&self) -> &[u32]

A flat view of the index buffer of this mesh.

pub fn reverse(&mut self)

Reverse the orientation of the triangle mesh.

pub fn triangles(&self) -> impl ExactSizeIterator<Item = Triangle> + '_

An iterator through all the triangles of this mesh.

impl TriMesh

pub fn flags(&self) -> TriMeshFlags

The flags of this triangle mesh.

pub fn aabb(&self, pos: &Isometry<f32>) -> Aabb

Compute the axis-aligned bounding box of this triangle mesh.

pub fn local_aabb(&self) -> &Aabb

Gets the local axis-aligned bounding box of this triangle mesh.

pub fn qbvh(&self) -> &Qbvh<u32>

The acceleration structure used by this triangle-mesh.

pub fn num_triangles(&self) -> usize

The number of triangles forming this mesh.

pub fn is_backface(&self, feature: FeatureId) -> bool

Does the given feature ID identify a backface of this trimesh?

pub fn triangle(&self, i: u32) -> Triangle

Get the i-th triangle of this mesh.

pub fn vertices(&self) -> &[Point<f32>]

The vertex buffer of this mesh.

pub fn indices(&self) -> &[[u32; 3]]

The index buffer of this mesh.

pub fn topology(&self) -> Option<&TriMeshTopology>

Returns the topology information of this trimesh, if it has been computed.

pub fn connected_components(&self) -> Option<&TriMeshConnectedComponents>

Returns the connected-component information of this trimesh, if it has been computed.

pub fn connected_component_meshes( &self, flags: TriMeshFlags, ) -> Option<Vec<TriMesh>>

Returns the connected-component of this mesh.

The connected-components are returned as a set of TriMesh build with the given flags.