parry2d::shape
Trait PolygonalFeatureMap

pub trait PolygonalFeatureMap: SupportMap {
    // Required method
    fn local_support_feature(
        &self,
        dir: &Unit<Vector<f32>>,
        out_feature: &mut PolygonalFeature,
    );

    // Provided method
    fn is_convex_polyhedron(&self) -> bool { ... }
}

Trait implemented by convex shapes with features with polyhedral approximations.

Required Methods

fn local_support_feature( &self, dir: &Unit<Vector<f32>>, out_feature: &mut PolygonalFeature, )

Compute the support polygonal face of self towards the dir.

Provided Methods

fn is_convex_polyhedron(&self) -> bool

Is this shape a ConvexPolyhedron?