Module parry2d::utils

Various unsorted geometrical and logical operators.

Modules

hashmap
A hash-map that behaves deterministically when the enhanced-determinism feature is enabled.

Structs

DeterministicState
A hasher builder that creates DefaultHasher with default keys.

HashablePartialEq
A structure that implements Eq and is hashable even if the wrapped data implements only PartialEq.

SdpMatrix2
A 2x2 symmetric-definite-positive matrix.

SdpMatrix3
A 3x3 symmetric-definite-positive matrix.

SortedPair
A pair of elements sorted in increasing order.

Traits

AsBytes
Trait that transforms thing to a slice of u8.

IsometryOps
Extra operations with isometries.

IsometryOpt
Various operations usable with Option<Isometry> and Option<&Isometry> where None is assumed to be equivalent to the identity.

Functions

ccw_face_normal
Computes the direction pointing toward the right-hand-side of an oriented segment.

center
Computes the center of a set of point.

cov
Computes the covariance matrix of a set of points.

median
Computes the median of a set of values.

point_cloud_support_point
Computes the support point of a cloud of points.

point_cloud_support_point_id
Computes the index of the support point of a cloud of points.

point_in_poly2d
Tests if the given point is inside of a polygon with arbitrary orientation.