parry2d::utils
Function point_in_poly2d

pub fn point_in_poly2d(pt: &Point2<f32>, poly: &[Point2<f32>]) -> bool

Tests if the given point is inside an arbitrary closed polygon with arbitrary orientation, using a counting winding strategy.

The polygon is assumed to be closed, i.e., first and last point of the polygon are implicitly assumed to be connected by an edge.

This handles concave polygons. For a function dedicated to convex polygons, see point_in_convex_poly2d.