Module parry2d::utils::point_in_triangle

Function to check if a point is inside a triangle and related functions.

Enums

Orientation

The orientation or winding direction of a corner or polygon.

Functions

corner_direction

Returns the direction of a line through p1, p2 and p3.

is_point_in_triangle

Returns true if point p is in triangle with corners v1, v2 and v3. Returns None if the triangle is invalid i.e. all points are the same or on a straight line.