Trait parry2d::utils::IntervalFunction

pub trait IntervalFunction<T> {
    // Required methods
    fn eval(&self, t: T) -> T;
    fn eval_interval(&self, t: Interval<T>) -> Interval<T>;
    fn eval_interval_gradient(&self, t: Interval<T>) -> Interval<T>;
}

A derivable valued function which can be bounded on intervals.

Required Methods

fn eval(&self, t: T) -> T

Evaluate the function at t.

fn eval_interval(&self, t: Interval<T>) -> Interval<T>

Bounds all the values of this function on the interval t.

fn eval_interval_gradient(&self, t: Interval<T>) -> Interval<T>

Bounds all the values of the gradient of this function on the interval t.