parry2d
Module math

Compilation flags dependent aliases for mathematical types.

Re-exports

pub use f32 as Real;

Constants

DEFAULT_EPSILON

The default tolerance used for geometric operations.

DIM

The dimension of the space.

SIMD_LAST_INDEX

SIMD_WIDTH - 1

SIMD_WIDTH

The number of lanes of a SIMD number.

TWO_DIM

The dimension of the space multiplied by two.

Type Aliases

AngDim

The dimension of the rotations.

AngVector

The angular vector type.

AngularInertia

The angular inertia of a rigid body.

CrossMatrix

A matrix that represent the cross product with a given vector.

Dim

The dimension of the ambient space.

Isometry

The transformation matrix type.

Matrix

The matrix type.

Orientation

The orientation type.

Point

The point type.

PrincipalAngularInertia

The principal angular inertia of a rigid body.

Rotation

The rotation matrix type.

SdpMatrix

A 2D symmetric-definite-positive matrix.

SimdBool

A SIMD bool with SIMD_WIDTH lanes.

SimdReal

A SIMD float with SIMD_WIDTH lanes.

SpacialVector

A vector with a dimension equal to the maximum number of degrees of freedom of a rigid body.

Translation

The translation type.

UnitVector

The unit vector type.

Vector

The vector type.