parry2d
Module partitioning

Spatial partitioning tools.

Structs

CenterDataSplitter

A data splitter that arranges a set of Aabbs in two sets based on their center’s coordinate along the split axis.

NodeIndex

The index of a node part of a Qbvh. The index of one specific node of a Qbvh.

Qbvh

A quaternary bounding-volume-hierarchy.

QbvhNode

A SIMD node of an SIMD Qbvh.

QbvhNonOverlappingDataSplitter

Data splitter for Qbvh construction that generates non-overlapping Aabbs at each level of the tree.

QbvhProxy

Combination of a leaf data and its associated node’s index.

QbvhUpdateWorkspace

Workspace for QBVH update.

Enums

SimdBestFirstVisitStatus

The next action to be taken by a BVH traversal algorithm after having visited a node with some data.

SimdSimultaneousVisitStatus

The status of the simultaneous traversal of two spatial partitioning structures.

SimdVisitStatus

The status of the spatial partitioning structure traversal.

Traits

IndexedData

A data to which an index is associated.

QbvhDataGenerator

Trait used for generating the content of the leaves of the Qbvh acceleration structure.

SimdBestFirstVisitor

Trait implemented by cost functions used by the best-first search on a BVT.

SimdSimultaneousVisitor

Trait implemented by visitor called during a simultaneous spatial partitioning data structure tarversal.

SimdVisitor

Trait implemented by visitor called during the traversal of a spatial partitioning data structure.

SimdVisitorWithContext

Trait implemented by visitor called during the traversal of a spatial partitioning data structure.

Type Aliases

SimdNodeIndex

The index of an internal SIMD node of a Qbvh.

SimdQbvhDeprecated

A quaternary bounding-volume-hierarchy.