parry2d::partitioning
Trait SimdBestFirstVisitor

pub trait SimdBestFirstVisitor<LeafData, SimdBV> {
    type Result;

    // Required method
    fn visit(
        &mut self,
        best_cost_so_far: f32,
        bv: &SimdBV,
        value: Option<[Option<&LeafData>; 4]>,
    ) -> SimdBestFirstVisitStatus<Self::Result>;
}

Trait implemented by cost functions used by the best-first search on a BVT.

Required Associated Types

type Result

The result of a best-first traversal.

Required Methods

fn visit( &mut self, best_cost_so_far: f32, bv: &SimdBV, value: Option<[Option<&LeafData>; 4]>, ) -> SimdBestFirstVisitStatus<Self::Result>

Compute the next action to be taken by the best-first-search after visiting a node containing the given bounding volume.