Module parry2d::query::point

Point inclusion and projection.

Structs

PointCompositeShapeProjBestFirstVisitor
A visitor for the projection of a point on a composite shape.

PointCompositeShapeProjWithFeatureBestFirstVisitor
A visitor for the projection of a point on a composite shape.

PointCompositeShapeProjWithLocationBestFirstVisitor
A visitor for the projection of a point on a composite shape.

PointProjection
Description of the projection of a point on a shape.

Traits

PointQuery
Trait of objects that can be tested for point inclusion and projection.

PointQueryWithLocation
Returns shape-specific info in addition to generic projection information

Functions

local_point_projection_on_support_map
Projects a point on a shape using the GJK algorithm.