parry2d
Module shape

Shapes supported by parry.

Structs

Ball

A Ball shape.

Capsule

A capsule shape defined as a round segment.

Compound

A compound shape with an aabb bounding volume.

ConvexPolygon

A 2D convex polygon.

Cuboid

Shape of a box.

HalfSpace

A half-space delimited by an infinite plane.

HeightField

A 2D heightfield with a generic storage buffer for its heights.

PackedFeatureId

A feature id where the feature type is packed into the same value as the feature index.

PolygonalFeature

A polygonal feature representing the local polygonal approximation of a vertex, or face, of a convex shape.

Polyline

A polyline.

RoundShape

A shape with rounded borders.

Segment

A segment shape.

SharedShape

The shape of a collider.

TopoFace

A face of a triangle-mesh’s half-edge topology.

TopoHalfEdge

A half-edge of a triangle-mesh’s half-edge topology.

TopoVertex

A vertex of a triangle-mesh’s half-edge topology.

TriMesh

A triangle mesh.

TriMeshConnectedComponents

The connected-components of a triangle mesh.

TriMeshFlags

The status of the cell of an heightfield.

TriMeshTopology

The half-edge topology information of a triangle mesh.

Triangle

A triangle shape.

TrianglePseudoNormals

The pseudo-normals of a triangle providing approximations of its feature’s normal cones.

Enums

FeatureId

An identifier of a feature of a convex polyhedron.

SegmentPointLocation

Logical description of the location of a point on a triangle.

ShapeType

Enum representing the type of a shape.

TopologyError

Indicated an inconsistency in the topology of a triangle mesh.

TriangleOrientation

Orientation of a triangle.

TrianglePointLocation

Description of the location of a point on a triangle.

TypedShape

Enum representing the shape with its actual type

Traits

PolygonalFeatureMap

Trait implemented by convex shapes with features with polyhedral approximations.

Shape

Trait implemented by shapes usable by Rapier.

SimdCompositeShape

Trait implemented by shapes composed of multiple simpler shapes.

SupportMap

Traits of convex shapes representable by a support mapping function.

TypedSimdCompositeShape

Type Aliases

HeightFieldCellStatus

Indicates if a cell of a heightfield is removed or not. Set this to false for a removed cell.

RoundConvexPolygon

A convex polygon dilated by a sphere (so it has round corners).

RoundCuboid

A cuboid dilated by a sphere (so it has round corners).

RoundTriangle

A triangle dilated by a sphere (so it has round corners).