parry2d::shape
Struct ConvexPolygon

pub struct ConvexPolygon { /* private fields */ }

A 2D convex polygon.

Implementations

impl ConvexPolygon

pub fn aabb(&self, pos: &Isometry<f32>) -> Aabb

Computes the world-space Aabb of this convex polygon, transformed by pos.

pub fn local_aabb(&self) -> Aabb

Computes the local-space Aabb of this convex polygon.

impl ConvexPolygon

pub fn bounding_sphere(&self, pos: &Isometry<f32>) -> BoundingSphere

Computes the world-space bounding sphere of this convex polygon, transformed by pos.

pub fn local_bounding_sphere(&self) -> BoundingSphere

Computes the local-space bounding sphere of this convex polygon.

impl ConvexPolygon

pub fn from_convex_hull(points: &[Point<f32>]) -> Option<Self>

Creates a new 2D convex polygon from an arbitrary set of points.

This explicitly computes the convex hull of the given set of points. Returns None if the convex hull computation failed.

pub fn from_convex_polyline(points: Vec<Point<f32>>) -> Option<Self>

Creates a new 2D convex polygon from a set of points assumed to describe a counter-clockwise convex polyline.

Convexity of the input polyline is not checked. Returns None if all points form an almost flat line.

pub fn points(&self) -> &[Point<f32>]

The vertices of this convex polygon.

pub fn normals(&self) -> &[Unit<Vector<f32>>]

The normals of the edges of this convex polygon.

pub fn scaled(self, scale: &Vector<f32>) -> Option<Self>

Computes a scaled version of this convex polygon.

Returns None if the result had degenerate normals (for example if the scaling factor along one axis is zero).

pub fn offsetted(&self, amount: f32) -> Self

Returns a mitered offset of the polygon.

Arguments

amount - size of the inflation. Each edge is moved outwards by this amount.

Panics

Panics if amount is not a non-negative finite number.

pub fn support_feature_id_toward( &self, local_dir: &Unit<Vector<f32>>, ) -> FeatureId

Get the ID of the feature with a normal that maximizes the dot product with local_dir.

pub fn feature_normal(&self, feature: FeatureId) -> Option<Unit<Vector<f32>>>

The normal of the given feature.