parry2d::shape
Trait SupportMap

pub trait SupportMap {
    // Required method
    fn local_support_point(&self, dir: &Vector<f32>) -> Point<f32>;

    // Provided methods
    fn local_support_point_toward(&self, dir: &Unit<Vector<f32>>) -> Point<f32> { ... }
    fn support_point(
        &self,
        transform: &Isometry<f32>,
        dir: &Vector<f32>,
    ) -> Point<f32> { ... }
    fn support_point_toward(
        &self,
        transform: &Isometry<f32>,
        dir: &Unit<Vector<f32>>,
    ) -> Point<f32> { ... }
}

Traits of convex shapes representable by a support mapping function.

Parameters:

V - type of the support mapping direction argument and of the returned point.

Required Methods

fn local_support_point(&self, dir: &Vector<f32>) -> Point<f32>

Provided Methods

fn local_support_point_toward(&self, dir: &Unit<Vector<f32>>) -> Point<f32>

Same as self.local_support_point except that dir is normalized.

fn support_point( &self, transform: &Isometry<f32>, dir: &Vector<f32>, ) -> Point<f32>

fn support_point_toward( &self, transform: &Isometry<f32>, dir: &Unit<Vector<f32>>, ) -> Point<f32>

Same as self.support_point except that dir is normalized.