parry2d
Module transformation

Transformation, simplification and decomposition of meshes.

Modules

utils

Utilities useful for various generations tasks.

vhacd

Approximate convex decomposition using the VHACD algorithm.

voxelization

Voxelization of a 2D polyline or 3D triangle mesh.

Functions

convex_hull

Computes the convex hull of a set of 2d points.

convex_hull_idx

Computes the convex hull of a set of 2d points and returns only the indices of the hull vertices.

convex_polygons_intersection

Computes the intersection of two convex polygons.

convex_polygons_intersection_points

Computes the intersection points of two convex polygons.

convex_polygons_intersection_points_with_tolerances

Computes the intersection points of two convex polygons.

convex_polygons_intersection_with_tolerances

Computes the intersection of two convex polygons.

hertel_mehlhorn

The Hertel-Mehlhorn algorithm.

hertel_mehlhorn_idx

Internal implementation of the Hertel-Mehlhorn algorithm that returns the polygon indices.

polygons_intersection

Compute intersections between two polygons that may be non-convex but that must not self-intersect.

polygons_intersection_points

Compute intersections between two polygons that may be non-convex but that must not self-intersect.